# docs/swagger.yaml
basePath: /
definitions:
...
  models.Age:
    properties:
      age:
        type: string
      id:
        type: integer
      posterUrl:
        type: string
    type: object
  models.AllSeries:
    properties:
      description:
        type: string
      director:
        type: string
      id:
        description: Указатель на int
        type: integer
      rating:
        type: integer
      release_year:
        type: integer
      series:
        type: integer
      title:
        type: string
      trailer_url:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Category:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.Movie:
    properties:
      ages:
        items:
          $ref: '#/definitions/models.Age'
        type: array
      allseries:
        items:
          $ref: '#/definitions/models.AllSeries'
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      description:
        type: string
      director:
        type: string
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      is_watched:
        type: boolean
      poster_url:
        type: string
      rating:
        type: integer
      release_year:
        type: integer
      title:
        type: string
      trailer_url:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: "OZINSHE\tAPI"
  version: "1.0"
paths:
...
  /movies:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of movies
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all movies
      tags:
      - movies
    post:
      consumes:
      - application/json
      parameters:
      - description: Title of the movie
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: Description of the movie
        in: body
        name: description
        required: true
        schema:
          type: string
      - description: ReleaseYear of the movie
        in: body
        name: releaseYear
        required: true
        schema:
          type: integer
      - description: Director
        in: body
        name: director
        required: true
        schema:
          type: string
      - description: TrailerUrl
        in: body
        name: trailerUrl
        required: true
        schema:
          type: string
      - description: Genre ids
        in: body
        name: genreIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: Category ids
        in: body
        name: categoryIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: Age ids
        in: body
        name: ageIds
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Could not bind json
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create movie
      tags:
      - movies
  /movies/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete movie
      tags:
      - movies
    get:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.Movie'
        "400":
          description: Invalid Movie Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Find by id
      tags:
      - movies
    put:
      consumes:
      - application/json
      parameters:
      - description: Title of the movie
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: Description of the movie
        in: body
        name: description
        required: true
        schema:
          type: string
      - description: ReleaseYear of the movie
        in: body
        name: releaseYear
        required: true
        schema:
          type: integer
      - description: Director
        in: body
        name: director
        required: true
        schema:
          type: string
      - description: TrailerUrl
        in: body
        name: trailerUrl
        required: true
        schema:
          type: string
      - description: Genre ids
        in: body
        name: genreIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: Category ids
        in: body
        name: categoryIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: Age ids
        in: body
        name: ageIds
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Could not bind json
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update movie
      tags:
      - movies
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"